AeHandler read expostime as long

+++ app/src/main/java/freed/cam/apis/camera2/parameters/AeHandler.java
@@ -282,12 +282,12 @@ public class AeHandler
                     millimax = 10000000;
                     break;
                 default:
-                    millimax = cameraHolder.characteristics.get(CameraCharacteristics.SENSOR_INFO_EXPOSURE_TIME_RANGE).getUpper().intValue() / 1000;
+                    millimax = (int)cameraHolder.characteristics.get(CameraCharacteristics.SENSOR_INFO_EXPOSURE_TIME_RANGE).getUpper().longValue() / 1000;
                     if (millimax == 0)
                         millimax = 800000;
                     break;
             }
-            int millimin = cameraHolder.characteristics.get(CameraCharacteristics.SENSOR_INFO_EXPOSURE_TIME_RANGE).getLower().intValue() / 1000;
+            int millimin = (int)cameraHolder.characteristics.get(CameraCharacteristics.SENSOR_INFO_EXPOSURE_TIME_RANGE).getLower().longValue() / 1000;
             stringvalues = getSupportedShutterValues(millimin, millimax,false);
         }
